Handle SuperBMD errors in a more user-friendly manner.
The exit code of SuperBMD was not being tested. When an error was
produced, the editor could end up loading the stale 3D model of the last
successful conversion (if one existed). If no previous 3D model existed,
then a cryptic error was produced ("No such file or directory"), without
really exposing to the user why the file is missing, or why it failed.

Also, since 35918ff988ba, the stale model could be wrongly cached for
the failed BMD model of the current custom track.

The fix consists of checking the exit code returned by SuperBMD to raise
an exception on error. The exception includes the error code, as well as
the error output produced by SuperBMD.

In addition, before attempting to load a new model or collision file,
the previous geometry is reset, so that stale polygons are no longer
rendered after failing to load a separate model.