Enabled custom context menu policy by default.
The custom context menu was supposed to be enabled in the top-down view
(unless the application is in object insertion mode), which is the
initial state of the application.

However, the custom context menu was not enabled on startup. Instead,
the user had to either press `Esc` to reset the mouse mode, which would
enable the custom context menu, or toggle between the 2D and 3D view
twice, which would also leave the custom context menu enabled.