Added depth attachment to the mono-sampled framebuffer.
This fixes a regression introduced as part of https://github.com/RenolY2/mkdd-track-editor/commit/07b0e77b9893c43f28fcd39faa7a5837d186e5f8.

The mono-sampled framebuffer, used when multisampling is enabled, did
not have a depth buffer assigned, meaning that all picking was performed
*without* depth testing.

Not very often do *identifiable* objects overlap, so this issue was not
very relevant, or noticeable.

The issue was only discovered after implementing the hover effect on the
gizmo, where it became obvious that the wrong manipulator handle was
being highlighted when multisampling was enabled (which is the default).