Renamed labels and variable names of previously unknown fields in the BOL header.
- **Unknown 1**: Split into **LOD Bias** and **Dummy Start Line**.
- **Unknown 3**: Renamed to **Shadow Opacity**.
- **Unknown 5**: Renamed to **Sky Follow**.
- **Unknown 6**: Removed from UI; always `0x00` in all stock BOL files.

When **LOD Bias** is on, the game seems to generate mipmaps differently,
resulting in a sharper image in some cases. Value is used from
`CrsData::isTexLODBiasOn()`.

**Dummy Start Line**, although named, remains unknown. It is not clear
what this flag implies. The name is borrowed from the function that
retrieves the value from the BOL file: `CrsData::isDummyStartLineOn()`.

**Sky Follow**, which is only enabled in Sherbet Land and Ending,
enables a different set of matrix operations on the skydome. The name is
borrowed from the `CrsData::isSkyFollow()` function.