Show combo box for camera types.
Previously-unknown byte has been combined with the camera type, which
is now read as an `unsigned short`, which is retrieved by
`CrsData::Camera::getKind()`.

The name of the cameras have been sourced from the symbols map, by
inspecting the places where `getKind()` was used before instantiating
each of the cameras.

Camera type `0x0008` remains unknown. It seems the logic uses "OutView"
as a fallback, but `8` is never really seen in the code.