Split the **Unknown** field in Enemy Points into **Drift Supplement** and **No Mushroom Zone**.
The unsigned short is now treated as two different bytes.

The first byte is read and processed by `RivalBodyCtrl::initDrift()`,
which uses it as a supplement for another value previously calculated.

The second byte is read and processed by
`RivalItemCtrl::useItem_Kinoko_Normal()`, used to skip the whole
function if the byte is not `0x00`.