Provide the `SideWidget` with a scroll area.
The scroll area is added to stop the `SideWidget` from dictating the
minimum width of its container (the splitter in the main window). If the
data editor of the given selection grows larger than the width of the
splitter, a scroll bar is now shown, instead of resizing the splitter
(and thus the viewport).

Bonus changes:

- Imports have been tidied up, removing unused imports, and avoiding to
  pollute the global namespace.
- Spacing between name label and main data editor is now based on font
  sized, as opposed to hard-coded values.
- Debugging prints have been removed.
- A number of other redundant lines, commented code, and unused
  functions have been removed.
- Font size is now determined based on a fraction of the initial font
  size, as opposed to hard-coding an arbitrary, fixed point size.
- Comment label has been fixed up (although its usefulness is arguable
  and could perhaps be removed entirely in the next iteration).