Ensure tool tip for **Object Type** is updated when it changes.
This was a regression introduced as part of #79. In order to avoid
rebuilding the object-specific widgets more than required, signals were
blocked, but that also stopped tool tips from being updated.

The signal blocker has been removed; instead, the current object type
name is stored, to only update the widgets when it changes.