Determine next enemy point in newly added respawn points accounting for the direction of the road.
Follow-up to #80, where functionality was introduced for automatically
selecting the closest enemy point as the next enemy point of the newly
added respawn points.

This revised algorithm attempts to select the next enemy point by
inferring the direction of the road, ensuring that the newly added
respawn point does not point backwards.

https://github.com/RenolY2/mkdd-track-editor/assets/1853278/15d8634b-ddf7-4b2f-a64d-b443ee6c7b29