Only update 3D coordinates in status bar if idling.
With large collision files (10k triangles), the `get_3d_coordinates()`
function can become expensive (~60 ms), which inevitable limits FPS if
executed on every mouse move event.

Now the 3D coordinates are updated if the user is not tumbling the view,
or applying a transform to the selected objects.

In the future, if `get_3d_coordinates()` gets optimized enough, this
change could be reverted.