Show correct next enemy point in Dolphin hook.
A number of offsets have been tweaked to load the position array from
the correct address. Address to the relevant data structure is now
retrieved using `KartCtrl::getKartEnemy()` as reference.

Previously, a different function (`KartCtrl::GetKartCenterPtr()`, whose
definition is just below `KartCtrl::getKartEnemy()` in NTSC-U) was used,
but the positions shown in the Editor did not match the next enemy
point; some other data was being represented [that is yet to be
determined].

**Before the fix**, the *supposed* next enemy points can be drawn behind
the kart's position:

![MKDD Track Editor - Next enemy points in Dolphin hook (before fix)](https://github.com/RenolY2/mkdd-track-editor/assets/1853278/4e6d4034-8f9b-4568-b9d2-f159f4e92653)

**After the fix**, next enemy points are always drawn ahead of the kart,
and are reliably updated as soon as karts reach the enemy point (i.e. as
soon as karts *touch* the enemy points' spheres):

![MKDD Track Editor - Next enemy points in Dolphin hook (after fix)](https://github.com/RenolY2/mkdd-track-editor/assets/1853278/753c9573-7421-424b-bc7b-5f7f3a0e112d)

NOTE: Clips taken on a modified Baby Park with fewer enemy points for
clarify.