Preserve routes when copy/pasting objects or cameras.
This was a regression in 016f5885dca, where objects and cameras started
storing route instances instead of route indexes.

The pickling mechanism did not work well in this case, as the pasted
objects/cameras would present a brand-new route instance that does not
correspond to any route in the route list.

Now, before pickling the selected objects or cameras into binary data,
route instances are temporarily converted to route indexes. During the
paste action, route indexes are converted to route instances based on
the current route list.