Preserve assigned cameras when copy/pasting areas.
This was a regression in 016f588, where areas started storing cameras
instances instead of camera indexes.

The pickling mechanism did not work well in this case, as the pasted
areas would present a brand-new camera instance that does not correspond
to any camera in the route list.

Now, before pickling the selected areas into binary data, camera
instances are temporarily converted to camera indexes. During the paste
action, camera indexes are converted to camera instances based on
the current camera list.