Relax ray and triangle collision to accept hits on edges.
When a ray was thrown from an edge of a triangle, the
`Collision.collide_ray()` function would not hit said triangle, as edges
were considered outside of the triangle.

This issue was surfaced while trying to ground objects that had been
previously snapped to edges: the object would ignore the triangle that
*almost* contained the object, and would be grounded to other triangles
much further below.

`Line.collide()` has been updated with the fix, too, although the
function is not currently used (superseded by the recent
reimplementation with Numba).