Accept `0.0` as a valid triangle collision distance.
This was a regression with the new Numba-base triangle collision
implementation, introduced in 0e7a34c37b1e.

`0.0` was inadvertently used to indicate "no collision", but that is a
valid value when the ray origin is contained in the triangle. Grounding
objects that were already grounded could result in objects being
grounded to a much higher or lower triangle, as the closest triangle was
wrongly disregarded.

`-1.0` is now used. (Note that Numba is strongly typed; returning `None`
or other rich type is not possible.)