Update collision only once when terrain types are toggled.
This is relevant when the **Show All** button is pressed when many
terrain types had been toggled off, implying that each type is going to
be toggled on.

Previously, when a terrain type was toggled on, an update event was
executed. If many types were toggled at once, it would cause a notable
delay.

Now the update is deferred, resulting in a single update after all items
have been toggled on.