Verify document state after selection in viewport.
Follow-up to e5c13a4fef0e832648d, where the Undo system now monitors
changes in the selection set.

In order to accurately detect changes in selection, the document state
needs to be re-verified after a selection occurs in the viewport. The
viewport selection is not fully synced with the mouse press event: it
can happen instants later, after the document state has been changed as
a result of the potentially-editing event (i.e. the mouse click event).

The issue was easily reproducible on Windows.