Reimplement infinite grid with more primitive GL operations.
The previous implementation, based on `GL_FOG`, could show different
results depending on the implementation of the `GL_NICEST` *hint*. In
some Linux systems, the hint triggers a per-fragment shading (as desired
for our use case) but, on Windows systems, the shading is generally done
per-vertex, which was not ideal.

In the new reimplementation, the infinite grid is achieved by painting
the grid with the same sky color; lines accumulating in the horizon give
the *false* impression of a fog. Similarly, the ground is painted with a
gradient that converges into the sky color.

<img src="https://github.com/RenolY2/mkdd-track-editor/assets/1853278/ec2b15b6-9368-43a8-96b4-56f569045610" alt="MKDD Track Editor - Infinite Grid Reimplementation" title="MKDD Track Editor - Infinite Grid Reimplementation" width="540"/>