Avoid deselecting top-level tree items as soon as they are selected.
When a tree item is selected, a sequence of signals are fired,
eventually invoking `action_update_info()`, that deselects all tree
items that do not have an associated selected object.

This prevents the top-level items (**Enemy Paths** etc.) from being
selected, as they would be immediately deselected.

The logic reponsible for this deselection has been moved to the method
where tree items are selected as a result of a click in the viewport,
which was original intention.