Migrate camera position when toggling between view modes.
The camera position is updated based on the target position of both
cameras, with the intention of keeping both the top-down camera and the
3D camera pointing at the same target.

The target is determined based on the ray collision from the camera
position and the "ground". The ground is determined based either on the
loaded collision model, if available, or on the enemy path points, if
available, defaulting to the XZ plane if neither are available.