Dynamically adjust far plane in 3D view based on distance to loaded model.
Previously far plane was hardcoded to `160000.0`, which is generally a
fair value for Mario Kart: Double Dash!!'s courses, but can still show
obvious geometry cuts in the viewport when the user moves the camera
far away.

To mitigate the effects, the far plane is now calculated based on the
distance to the center of the loaded model (if present), otherwise
defaulting to the previous hardcoded value.