Change color of enemy path points for better differentiation from other types.
The new color is a tone of purple; previously, gray was used.

Rationale for the change: Unassigned route points, and snapping points
already use light gray and white colors. Also, road-like terrain types
are rendered in gray-ish colors as well; with enemy path points usually
landing in the middle of the road, a purple color is more recognizable.