Add custom crosshair cursor for insertion mode.
This is the cursor that is used when insertion mode (for placing
objects) is entered.

Previously, the cursor provided by the system/Qt was used, but the look
and feel could be substantially different between platforms.

The SVG image is rasterized to 32x32, as that's the resolution
recommended by Qt to support most platforms. Screens with greater pixel
density could welcome a greater resolution (e.g. 48x48)... To be
revisited in the future.