Attempt to load `.arc` file under `try`/`except` block.
When a file was loaded from the recently loaded list, an exception could
be raised if the file was already gone, without any visual clue in the
graphica user interface; only in the terminal/console.

Now files are loaded from a relevant `try`/`except` block that will
raise a popup dialog on error.