Ensure tool tip for **Object Type** is set on first show event.
This was a change in 7178d896693f4216b, where signals in the combo box
are now temporarily blocked, preventing the widget from setting the tool
tip unless the user changes the object type.

Bonus: "Description not available." is now shown for object types that
do not have a description yet.