Disallow manual input of duplicate IDs in enemy paths.
The **Group ID** field in the data editor was not checking whether the
value entered by the user was unique.

Furthermore, it was not updating the group ID in the child enemy path
points, leaving the BOL document in an inconsistent state (at least
temporarily).

Bonus: The tree view item is now updated as the ID changes in the data
editor.