Hide scale Z component field in cylindrical areas.
The game (see `Course::Area::check()`) only checks the squared Euclidean
distance using the squared X component, disregarding the Z component.

The field is now hidden in the data editor when the area shape is set
to `Cylinder`, and the Z component is also ignored when the area is
drawn in the viewport.