Capture and show error messages produced while trying to find the Dolphin instance.
Previously, a general "Dolphin not found." message was shown in the
error popup; rest of the potential error messages were only printed to
the console/terminal.

This was an inconvenience on Linux, as the error message was critical to
let the user know that the Dolphin Hook server needed to be running with
elevated permissions.