Explicitly provide `ClearAndSelect` flag to `QTreeWidget.setCurrentItem()`."
Certain keyboard modifiers (e.g. `Ctrl`) may make it behave differently
or erratically on occasions when the flag is not set.

This was fist noticed as part of the upcoming work to support `Shit` and
`Ctrl` modifiers for appending, toggling, or removing the current items
in the selection in the viewer.