Revise keyboard shortcuts in framing functionality.
Previously, `F` was used as keyboard shortcut for either framing the
current selection, or framing all objects (and loaded geometry) if there
is no selection.

To match the same functionality in Blender; most used modelling software
in the community; the `Numpad .` and `Home` keyboard shortcuts are now
handled too. For legacy reasons, the `F` keyboard shortcut is still
supported for framing selection objects:

- **Frame Selected      `Numpad . or F`**
- **Frame All           `Home`**

In order to limit the action of the `.` key to the set in the numeric
pad, the keyboard shortcut must be processed manually in the
`keyPressEvent()` handler.