Update rotation widgets in data editor on change.
When the selected object is rotated in the viewport, the rotation fields
in the data editor should now be updated interactively.

This was a regression in 6f166010c72c602, where the object's rotation
was wrongly cached when the widgets are created, totally overlooking
that any potential update coming from the rotate manipulators in the
viewport would be disregarded if the rotation is cached!