Add GitHub Actions to generate build on Linux (x64) and Windows (x64).
The preexisting `setup.py` script is used in the YAML build scripts to
generate the build. The actions will be triggered by:

- Any commit pushed to the main branch (`master` currently).
- A new `vx.y.z` tag pushed to the repository.

The builds will be compressed in ZIP files, with the following naming
format: `<app_name>-<version>-<sha>-<platform>-<architecture>`

For example:

- `mkdd-track-editor-1.4.0-d2f94af5-linux-x64.zip`
- `mkdd-track-editor-1.4.0-d2f94af5-windows-x64.zip`

For actions triggered by a tag creation, the commit SHA will be omitted.