Support rotations around X axis and Z axis in **Rotate Around Median Point** mode.
When the **Rotate Around Median Point** mode is on, objects rotate
around the median point (i.e. the position of the gizmo) as opposed to
rotating around their individual origins.

Previously, only rotations around the Y axis were supported in this
mode. Now, this is handled also for rotations around the X and Z axes.