Cache BMD models with an atomic operation.
Previously, the resulting converted folder by SuperBMD was copied to the
final destination directly. However, if the operation was to be
interrupted, the cache entry could be invalid.

Now the converted folder is first copied to an alternative directory;
only when the operation is complete is the folder renamed to its
final destination, which the system guarantees is an atomic operation.