Add **Apply Current View to Selected Object** action.
![MKDD Track Editor - Apply Current View to Selected Object](https://github.com/user-attachments/assets/2a2c2e2f-bf84-4b12-ad5c-fa01b4760e70)

The action is available if the 3D View is on and there is at least one
*positioned* object selected (i.e. an object that features a position
property and a rotation property). If there is more than one object
selected, the view is applied only to the last object.

This is useful for adjusting the position of static cameras by first
moving the 3D view to the wanted position and then transferring the view
to the selected camera.